requirements_test.txt
# linters such as flake8 and pylint should be pinned, as new releases
# make new things fail. Manually update these pins when pulling in a
# new version

# types-* that have versions roughly corresponding to the packages they
# contain hints for available should be kept in sync with them

-r requirements.txt
codecov==2.1.12
coverage==6.3.3
freezegun==1.2.1
mock-open==1.4.0
mypy==0.950
pipdeptree==2.2.1
pylint-strict-informational==0.1
pytest-aiohttp==0.3.0
pytest-cov==3.0.0
pytest-freezegun==0.4.2
pytest-socket==0.5.1
pytest-test-groups==1.0.3
pytest-sugar==0.9.4
pytest-timeout==2.1.0
pytest-xdist==2.5.0
pytest==7.1.1
requests_mock==1.9.2
respx==0.19.0
stdlib-list==0.7.0
tqdm==4.49.0
types-atomicwrites==1.4.1
types-croniter==1.0.0
types-backports==0.1.3
types-certifi==0.1.4
types-chardet==0.1.5
types-decorator==0.1.7
types-enum34==0.1.8
types-ipaddress==0.1.5
types-pkg-resources==0.1.3
types-python-slugify==0.1.2
types-pytz==2021.1.2
types-PyYAML==5.4.6
types-requests==2.25.1
types-toml==0.1.5
types-ujson==0.1.1
